Move CurrencyFormatter to fiat
